requirements.txt
solana==0.30.2
solders>=0.18.0
datetime
python-dotenv==1.0.0
schedule==1.2.0
pytz==2023.3
anthropic==0.3.11
qiskit==0.44.1
qiskit-aer==0.12.2
numpy>=1.22.0
